gingacc-player (nclua): Fix bug when sending "attribution" event
* src/application/imperative/lua/LuaPlayer.cpp
(setPropertyValue): We were sending an event with type="presentation"
instead of "attribution", which is the correct one.  This commit fixes
this bug.
(post, setPropertyValue, userEventReceived): Update nclua_sendx calls
to match the new prototype.
* src/application/imperative/lua/nclua.h (nclua_send, nclua_sendx):
Remove "index" parameter from the prototype.  Now these functions
assume that the event is at top of stack.
* src/application/imperative/lua/nclua-event-deprecated.cpp
(nclua_sendx): Likewise.  The function now assumes that the event is
at top of stack and pops it before returning.
* src/application/imperative/lua/nclua.c (nclua_destroy): Fix segfault
when releasing user data.
* src/application/imperative/lua/nclua-event.cpp (notify): Cleanup
stack before returning.