gingacc-player (nclua): Remove arbitrary limits in LuaCanvas.cpp
This commit also updates the font load policy.  First, we try to load
the font from the path given.  If this path is invalid, we search for
the font in the default directory.  Finally, if the font is not there,
we use the default font (currently "vera.ttf").

* src/application/imperative/lua/LuaCanvas.cpp
(Canvas): Use malloc-ed string for face name and style name.
(l_attrFont): Updated accordingly.
(l_gc): Release the malloc-ed strings.
(l_createcanvas): Cleanup canvas' fields.